🚚 [RUMF-848] remove tracekit IIFE by spliting it into different modules
