+++ packages/core/src/domain/tracekit/index.ts
@@ -1,0 +1,3 @@
+export { Handler, StackTrace, BrowserError } from './types'
+export { computeStackTrace } from './computeStackTrace'
+export { report } from './report'
